==> 01_data_prep_for_simple_model.ipynb <==
# Data cleaning for simple model

import cryptpandas as crp
import datetime as dt
import numpy as np
import pandas as pd

from IPython.display import clear_output
from utils.missing import replace_missing_values

## Load files

impute_method = pd.read_csv('./data/missing_values_method.csv')

data_info = pd.read_csv('./data/SMMIS_fields.csv')
data_info.head()

password = input('Password? ')
clear_output()

filename = './data/SMMIS data (1988-2000).crypt'
data = crp.read_encrypted(filename, password=password)
data.set_index('recordno', inplace=True)

# Save unecrypted for inspection
sample = data.sample(100).T
sample.to_csv('./data/SMMIS_sample.csv')

# Remove data with no 5 min Apgar score
mask = data['apgar5'].notnull()
data = data[mask]

Filter data to fileds used in base model

mask = data_info['Base model'] == 1
fields = list(data_info['Variable name'][mask])
fields = [field.lower() for field in fields]
data = data[fields]
data.info()

# Format data values
cols = ['mdob','lperiod', 'birthdat']

for col in cols:
    data[col] = pd.to_datetime(data[col])

## Replace missing values for selected fields

# Replace missing induction method with 'NA'
data['methind'] = replace_missing_values(data['methind'], 'NA')

# Replace missing hight and weight with zero
data['mweight'] = replace_missing_values(data['mweight'], 'zero')
data['mheight'] = replace_missing_values(data['mheight'], 'zero')

# Keep complete data only
mask = data.isnull().sum(axis=1) == 0
data = data[mask]
data.info()

## Calculate required fields

# Calculate BMI
bmi = data['mweight'] / ((data['mheight']/100) ** 2)
bmi =  bmi.values.astype(int)

# Replace any BMI < 10 with zero
mask = bmi < 10
bmi[mask]= 0

data['bmi'] = bmi
data.drop(['mweight', 'mheight'], axis=1, inplace=True)



# Calculate age (years)
age = data['birthdat'] - data['mdob']
age = age.dt.days / 365
data ['age'] = age.values.astype(int)

# Calculate gestantion age (weeks)
gest = data['birthdat'] - data['lperiod']
gest = gest.dt.days / 7
data['gest'] = gest.values.astype(int)

# Remove dates
data.drop(['birthdat', 'mdob', 'lperiod'], axis=1, inplace=True)

# Remove any gestatton <20 weeks or >45 weeks
mask = (data['gest'] <= 45) & (data['gest'] >= 20)
data = data[mask]

# Remove parity > 10
mask = data['parity'] <= 10
data = data[mask]


data.info()

# Remove subzero age and gestational age
mask = (data['age'] > 0) & (data['gest'] > 0)
data = data[mask]

# Reduce ethnicity
data['ethnic_white'] = data['ethnic'] == 'C'
data.drop('ethnic', axis=1, inplace=True)

# Randomise
data = data.sample(frac=1.0)

data.describe()

data.head().T

data.to_csv('./data/data_for_model.csv', index=False)
